dash board rework progress.

+++ PTO-WEB/src/main/java/com/thorton/grant/uspto/prototypewebapp/controllers/ApplicationController.java
@@ -32,7 +32,7 @@ public class ApplicationController {
         this.serviceBeanFactory = serviceBeanFactory;
     }
 
-    private static int counter = 0;
+    private static long counter = 0;
 
 
     private boolean continuation = false;
